Split the history containers for snapshot/hotline
[#587]
