+++ app/javascript/containers/snapshot/HistoryTableContainer.js
@@ -3,7 +3,6 @@ import HistoryTable from 'views/history/HistoryTable'
 import {
   getFormattedCasesSelector,
   getFormattedReferralsSelector,
-  getFormattedScreeningsSelector,
 } from 'selectors/screening/historyOfInvolvementSelectors'
 import * as IntakeConfig from 'common/config'
 
@@ -11,7 +10,7 @@ const mapStateToProps = (state) => ({
   showCopyButton: IntakeConfig.jsClipboardSupported(),
   cases: getFormattedCasesSelector(state).toJS(),
   referrals: getFormattedReferralsSelector(state).toJS(),
-  screenings: getFormattedScreeningsSelector(state).toJS(),
+  screenings: [],
   // To make the copied table fit in MS Word, we have to temporarily resize it.
   onCopy: (copyContent) => {
     copyContent.style.width = '1%'
@@ -26,4 +25,3 @@ const mapStateToProps = (state) => ({
 })
 
 export default connect(mapStateToProps)(HistoryTable)
-
